SonarQube fixes: * Lint fixes from master which for some reason didn't fail the build * Rename files based on their default export * Fix audit/security warnings
