+++ examples/complete/car-models/redirect.html
@@ -1,12 +1,11 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
+    <title>Redirect</title>
     <meta charset="UTF-8">
 </head>
 <body>
-This page should redirect to <a href="http://localhost:8080">http://localhost:8080</a>
-<script>
-    location.href = "http://localhost:8080";
-</script>
+    This page should redirect to <a href="http://localhost:8080">http://localhost:8080</a>
+    <script>location.href = "http://localhost:8080";</script>
 </body>
 </html>
