+++ examples/complete/car-models/src/app.html
@@ -3,7 +3,7 @@
 </div>
 <div class="container"  ng-if="$ctrl.connected">
 	<h1>Car models</h1>
-	<p>Comparison of <b>Horse power</b> across different car models.</p>
+	<p>Comparison of <strong>Horse power</strong> across different car models.</p>
 	<div style="margin-bottom: 10px;" ng-if="$ctrl.painted">
 		<a href="#" ng-click="$ctrl.sortByModel()">Sort by model</a>
 		&nbsp;&nbsp;
