+++ package.json
@@ -41,6 +41,7 @@
     "@babel/core": "7.5.5",
     "@babel/plugin-transform-modules-commonjs": "7.5.0",
     "@babel/preset-env": "7.5.5",
+    "@joseph184/rollup-plugin-node-builtins": "2.1.4",
     "babel-plugin-istanbul": "5.2.0",
     "coveralls": "3.0.6",
     "cross-env": "5.2.0",
@@ -57,7 +58,6 @@
     "rollup-plugin-commonjs": "10.0.2",
     "rollup-plugin-filesize": "6.2.0",
     "rollup-plugin-license": "0.12.1",
-    "rollup-plugin-node-builtins": "2.1.2",
     "rollup-plugin-node-globals": "1.4.0",
     "rollup-plugin-node-resolve": "5.2.0",
     "rollup-plugin-uglify": "6.0.2",
