+++ rollup.config.js
@@ -1,7 +1,7 @@
 // rollup.config.js
 import resolve from 'rollup-plugin-node-resolve';
 // import nodeGlobals from 'rollup-plugin-node-globals';
-import nodeBuiltins from 'rollup-plugin-node-builtins';
+import nodeBuiltins from '@joseph184/rollup-plugin-node-builtins';
 import commonjs from 'rollup-plugin-commonjs';
 import babel from 'rollup-plugin-babel';
 import { uglify } from 'rollup-plugin-uglify';
