+++ src/interceptors/delta-request-interceptor.js
