+++ test/unit/cache.spec.js
@@ -1,4 +1,4 @@
-import KeyValueCache from '../../src/cache';
+import KeyValueCache from '../../src/key-value-cache';
 
 describe('KeyValueCache', () => {
   let keyValueCache;
